dummy-data-product/src/requirements.txt
async-generator==1.10
attrs==22.1.0
Automat==20.2.0
beautifulsoup4==4.11.1
certifi==2022.6.15.1
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
constantly==15.1.0
cryptography==38.0.1
cssselect==1.1.0
decorator==5.1.1
filelock==3.8.0
future==0.18.2
geocoder==1.38.1
geographiclib==1.52
geopy==2.2.0
h11==0.13.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
itemadapter==0.7.0
itemloaders==1.0.6
jmespath==1.0.1
lxml==4.9.1
metadata-parser==0.10.5
numpy==1.23.3
outcome==1.2.0
pandas==1.4.4
parsel==1.6.0
Protego==0.2.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDispatcher==2.0.6
pyOpenSSL==22.0.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.2.1
PyVirtualDisplay==3.0
queuelib==1.6.2
ratelim==0.1.6
requests==2.28.1
requests-file==1.5.1
requests-toolbelt==0.9.1
Scrapy==2.6.2
selenium==4.4.3
service-identity==21.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
tldextract==3.3.1
tqdm==4.64.1
trio==0.21.0
trio-websocket==0.9.2
Twisted==22.8.0
typing-extensions==4.3.0
urllib3==1.26.12
w3lib==2.0.1
webdriver-manager==3.8.3
wsproto==1.2.0
zope.interface==5.4.0
